Fixed two bugs, one in riffa.v, one in tx_data_fifo.v
In riffa.v, the interrupt module was getting data from the TXC_DATA, which didn't match the assertion of the ready signal from the registers module. Changed to _wTxcData.

In tx_data_fifo.v, changed the DATA_READY signal to connect to the fifo instead of the packet counter.

In the tx_engine, removed the curious case of the +1 in the data fifo packet depth